Lets try use ptrdiff_t for pointer arithmetic

+++ PicoVM/CLR/AssemblyReader.hxx
@@ -2,6 +2,7 @@
 #define ASSEMBLYREADER_HXX
 #include <vector>
 #include <cstdint>
+#include <cstddef>
 #include <algorithm>
 
 #include "utf8.h"
@@ -78,14 +79,14 @@ class AssemblyReader {
 				static_cast<uint64_t>(*(it++)) << 56;
 		}
 
-		uint32_t read_asciiz(std::vector<uint8_t>& result, uint32_t limit) {
-			uint32_t offset = distance(data.begin(), pc);
-			uint32_t read = read_asciiz(result, offset, limit);
+		ptrdiff_t read_asciiz(std::vector<uint8_t>& result, uint32_t limit) {
+			ptrdiff_t offset = distance(data.begin(), pc);
+			ptrdiff_t read = read_asciiz(result, offset, limit);
 			pc = next(pc, read);
 			return read;
 		}
 
-        uint32_t read_asciiz(std::vector<uint8_t>& result, uint32_t offset, uint32_t limit) const {
+        ptrdiff_t read_asciiz(std::vector<uint8_t>& result, ptrdiff_t offset, uint32_t limit) const {
 			auto start_it = next(data.cbegin(), offset);
 			auto end_it = std::find(start_it, data.cend(), 0);
 			if (distance(start_it, end_it) >= limit) {
@@ -96,14 +97,14 @@ class AssemblyReader {
             return distance(start_it, end_it);
         }
 
-        uint32_t read_utf8z(std::vector<uint16_t>& result, uint32_t limit) {
-			uint32_t offset = distance(data.begin(), pc);
-			uint32_t read = read_utf8z(result, offset, limit);
+        ptrdiff_t read_utf8z(std::vector<uint16_t>& result, uint32_t limit) {
+			ptrdiff_t offset = distance(data.begin(), pc);
+			ptrdiff_t read = read_utf8z(result, offset, limit);
 			pc = next(pc, read);
 			return read;
 		}
 
-		uint32_t read_utf8z(std::vector<uint16_t>& result, uint32_t offset, uint32_t limit) const {
+		ptrdiff_t read_utf8z(std::vector<uint16_t>& result, ptrdiff_t offset, uint32_t limit) const {
 			auto start_it = next(data.cbegin(), offset);
 			auto end_it = std::find(start_it, data.cend(), 0);
 			if (distance(start_it, end_it) >= limit) {
